Fix ubuntu vagrant box nodesets
Without this patch, the nodesets incorrectly define a vagrantcloud vm
to install, resulting in the following error:

 > The box you're adding has a name different from the name you
 > requested.

This patch uses the trusty vagrant nodeset from the puppetlabs-apache
module and updates the default nodeset as well. Fill do the same for
the centos nodeset in a separate patch after RedHat tests are
implemented.

Change-Id: Iffac76b48d496d93b9cb63323ac103619319de47